# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/organization_statuses": {
            "get": {
                "description": "Get Organization Statuses\nGet a list of status codes for which organizations exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization Statuses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/organizations": {
            "get": {
                "description": "List Organizations\nList Organizations, paging with reverse, limit, and offset. Optionally, filter by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "PENDING",
                            "ENABLED",
                            "DISABLED"
                        ],
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/org.Organization"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Organization\nCreate a new Organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Organization",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/org.Organization"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Newly-created Organization",
                        "schema": {
                            "$ref": "#/definitions/org.Organization"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the newly created Organization"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid JSON body)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "422": {
                        "description": "Organization validation errors",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
...
        "/v1/organizations/{id}/versions": {
            "get": {
                "description": "Get Organization Versions\nGet Organization Versions by ID, paging with reverse, limit, and offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization Versions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/org.Organization"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/organizations/{id}/versions/{versionid}": {
            "get": {
                "description": "Get Organization Version\nGet Organization Version by ID and VersionID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization VersionID",
                        "name": "versionid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization Version",
                        "schema": {
                            "$ref": "#/definitions/org.Organization"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "head": {
                "description": "Organization Version Exists\nCheck if the specified Organization version exists.",
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization VersionID",
                        "name": "versionid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Organization Version Exists"
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter)"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.APIEvent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventID": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "org.Organization": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "versionID": {
                    "type": "string"
                }
            }
        },
...
    }
}